Feat: CGI 스태틱 클래스로 전환
